[DF-5024] Reverted File::Write and File::Seek to 32 bit implementation. Downloader adapted to 32 bit Save operations.
Former-commit-id: 9a56c4657af39056280bd2f1b2bd57db4faaa225

+++ Sources/Internal/DLC/Downloader/DownloaderCommon.cpp
@@ -106,7 +106,7 @@ bool DownloadPart::SaveDownload(const FilePath &infoFilePath)
     if (NULL != static_cast<File*>(file))
     {
         // we should to write pert data in it's place inside info file
-        uint64 seekPos = sizeof(DownloadInfoHeader) + info.number*sizeof(DownloadPart::StoreData);
+        uint32 seekPos = sizeof(DownloadInfoHeader) + info.number*sizeof(DownloadPart::StoreData);
         file->Seek(seekPos, File::SEEK_FROM_START);
 
         return sizeof(info) == file->Write(&info, sizeof(info));
